Refactored Item related utils package
