+++ src/main/java/me/zeroeightsix/kami/manager/managers/PlayerInventoryManager.kt
@@ -5,6 +5,8 @@ import me.zeroeightsix.kami.event.events.RenderOverlayEvent
 import me.zeroeightsix.kami.manager.Manager
 import me.zeroeightsix.kami.module.Module
 import me.zeroeightsix.kami.util.*
+import me.zeroeightsix.kami.util.items.clickSlot
+import me.zeroeightsix.kami.util.items.removeHoldingItem
 import me.zeroeightsix.kami.util.threads.safeListener
 import net.minecraft.client.gui.inventory.GuiContainer
 import net.minecraft.inventory.ClickType
