+++ src/main/java/me/zeroeightsix/kami/module/modules/combat/AutoOffhand.kt
@@ -10,6 +10,7 @@ import me.zeroeightsix.kami.util.combat.CombatUtils
 import me.zeroeightsix.kami.util.combat.CombatUtils.calcDamageFromMob
 import me.zeroeightsix.kami.util.combat.CombatUtils.calcDamageFromPlayer
 import me.zeroeightsix.kami.util.combat.CrystalUtils.calcCrystalDamage
+import me.zeroeightsix.kami.util.items.*
 import me.zeroeightsix.kami.util.text.MessageSendHelper
 import me.zeroeightsix.kami.util.threads.safeListener
 import net.minecraft.client.gui.inventory.GuiContainer
