+++ src/main/java/me/zeroeightsix/kami/module/modules/combat/CrystalAura.kt
@@ -19,6 +19,7 @@ import me.zeroeightsix.kami.util.combat.CrystalUtils.calcCrystalDamage
 import me.zeroeightsix.kami.util.combat.CrystalUtils.canPlaceCollide
 import me.zeroeightsix.kami.util.combat.CrystalUtils.getCrystalBB
 import me.zeroeightsix.kami.util.combat.CrystalUtils.getCrystalList
+import me.zeroeightsix.kami.util.items.*
 import me.zeroeightsix.kami.util.math.RotationUtils
 import me.zeroeightsix.kami.util.math.VectorUtils.distanceTo
 import me.zeroeightsix.kami.util.math.VectorUtils.toBlockPos
